+++ craftr/lib/craftr.ext.platform.py
@@ -36,7 +36,7 @@ Contents
 --------
 '''
 
-__all__ = ['WIN32', 'DARWIN', 'LINUX', 'CYGWIN', 'cc', 'cxx', 'ld', 'ar']
+__all__ = ['WIN32', 'DARWIN', 'LINUX', 'CYGWIN', 'as,', 'cc', 'cxx', 'ld', 'ar']
 
 from craftr import import_module
 from craftr.magic import Proxy
@@ -51,6 +51,7 @@ CYGWIN = 'cygwin' #: Cygwin platform name
 import_module(__name__ + '.' + sys.platform, globals(), '*')
 
 # get_tool() imported from the current platform module.
+asm = Proxy(get_tool, 'asm')  #: The Assembler retrieved with :func:`platform.get_tool`
 cc = Proxy(get_tool, 'cc')    #: The C compiler retrieved with :func:`platform.get_tool`
 cxx = Proxy(get_tool, 'cxx')  #: The C++ compiler retrieved with :func:`platform.get_tool`
 ld = Proxy(get_tool, 'ld')    #: The linker retrieved with :func:`platform.get_tool`